Move trigonometry implementation to its own file.
